refactor problem count function
